+++ config/services.php
@@ -1,6 +1,7 @@
 <?php
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
+use Psys\OrderInvoiceBundle\Twig\AppExtension;
 use Psys\OrderInvoiceBundle\Command\InstallCommand;
 use Psys\OrderInvoiceBundle\Command\StylerEnableCommand;
 use Psys\OrderInvoiceBundle\Command\Upgrade13To14Command;
@@ -17,6 +18,7 @@ use Psys\OrderInvoiceBundle\Service\FileDeleter\FileDeleter;
 use Psys\OrderInvoiceBundle\Service\FilePersister\FilePersister;
 use Psys\OrderInvoiceBundle\Service\InvoiceManager\InvoiceManager;
 use Psys\OrderInvoiceBundle\Service\InvoiceGenerator\MpdfGenerator;
+use Psys\OrderInvoiceBundle\Twig\OrderRuntime;
 use Psys\Utils\Math;
 
 return function(ContainerConfigurator $container): void 
@@ -67,11 +69,19 @@ return function(ContainerConfigurator $container): void
                 param('oi.storage_path')
             ])
             ->alias(FileDeleter::class, 'oi.file_deleter')
-
+        
         ->set(DoctrineSubscriber::class)
-        ->tag('doctrine.event_listener', 
+            ->tag('doctrine.event_listener', 
             [
                 'event' => 'onFlush',
+            ])
+
+        ->set(AppExtension::class)
+            ->tag('twig.extension')
+        ->set(OrderRuntime::class)
+            ->tag('twig.runtime')
+            ->args([
+                service('oi.order_manager'),
             ])
     ;