+++ src/EventSubscriber/DoctrineSubscriber.php
@@ -3,7 +3,8 @@ namespace Psys\OrderInvoiceBundle\EventSubscriber;
 
 use Doctrine\ORM\Event\OnFlushEventArgs;
 use Psys\OrderInvoiceBundle\Entity\Invoice;
-
+use Psys\OrderInvoiceBundle\Entity\InvoiceFinal;
+use Psys\OrderInvoiceBundle\Exception\InvalidInvoiceStateException;
 
 class DoctrineSubscriber
 {
@@ -14,11 +15,14 @@ class DoctrineSubscriber
 
         $ent_Invoice_insert = null;
         $ent_Invoice_update = null;
+        $ent_InvoiceFinal_insert = null;
 
         foreach ($uow->getScheduledEntityInsertions() as $entInsert) 
         {
             /** @var Invoice $ent_Invoice_insert  */
             if ($entInsert instanceof Invoice) {$ent_Invoice_insert = $entInsert;}
+            /** @var InvoiceFinal $ent_InvoiceFinal_insert  */
+            if ($entInsert instanceof InvoiceFinal) {$ent_InvoiceFinal_insert = $entInsert;}
         }
 
         foreach ($uow->getScheduledEntityUpdates() as $entUpdate) 
@@ -36,6 +40,20 @@ class DoctrineSubscriber
         {
             $this->checkInvoice($ent_Invoice_update);
         }
+
+        // If the final invoice is just being issued
+        if ($ent_InvoiceFinal_insert) 
+        {
+            // and at least one advance invoice exists
+            if (!$ent_Invoice_update->getInvoicesAdvance()->isEmpty())
+            {
+                // and order has no items
+                if ($ent_Invoice_update->getOrder()->getItems()->isEmpty())
+                {
+                    throw new InvalidInvoiceStateException('You need to add at least one item to the order (which represents the total value) from which advance invoices will be deducted.');
+                }
+            }
+        }
     }
 
     private function checkInvoice(Invoice $ent_Invoice): void
@@ -46,12 +64,12 @@ class DoctrineSubscriber
 
         if (empty($ent_InvoiceProforma) && $ent_InvoicesAdvance->isEmpty() && !empty($ent_InvoiceFinal)) 
         {
-            throw new \RuntimeException('Final invoice requires proforma or advance invoice to be issued first.');
+            throw new InvalidInvoiceStateException('Final invoice requires proforma or advance invoice to be issued first.');
         }
 
         if (!empty($ent_InvoiceProforma) && !$ent_InvoicesAdvance->isEmpty()) 
         {
-            throw new \RuntimeException('Proforma and advance invoice cannot be issued simultaneously.');
+            throw new InvalidInvoiceStateException('Proforma and advance invoice cannot be issued simultaneously.');
         }
     }