+++ src/Exception/InvalidInvoiceStateException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Psys\OrderInvoiceBundle\Exception;
+
+use DomainException;
+
+final class InvalidInvoiceStateException extends DomainException
+{
+}