+++ src/Twig/AppExtension.php
@@ -1,0 +1,16 @@
+<?php
+namespace Psys\OrderInvoiceBundle\Twig;
+
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+
+
+class AppExtension extends AbstractExtension
+{
+    public function getFilters(): array
+    {
+        return [
+            new TwigFilter('invoices_advance_totals', [OrderRuntime::class, 'getInvoicesAdvanceTotals']),
+        ];
+    }
+}